+++ packages/ckeditor5-table/tests/table-integration.js
@@ -1,6 +1,6 @@
 /**
- * @license Copyright (c) 2003-2018, CKSource - Frederico Knabben. All rights reserved.
- * For licensing, see LICENSE.md.
+ * @license Copyright (c) 2003-2019, CKSource - Frederico Knabben. All rights reserved.
+ * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
  */
 
 import Widget from '@ckeditor/ckeditor5-widget/src/widget';
@@ -95,37 +95,37 @@ describe( 'Table feature – integration', () => {
 		} );
 
 		it( 'fixing empty roots should be transparent to undo', () => {
-			expect( editor.getData() ).to.equal( '<p>&nbsp;</p>' );
+			expect( editor.getData( { trim: 'none' } ) ).to.equal( '<p>&nbsp;</p>' );
 			expect( editor.commands.get( 'undo' ).isEnabled ).to.be.false;
 
 			editor.data.set( viewTable( [ [ 'foo' ] ] ) );
 
-			expect( editor.getData() ).to.equal( viewTable( [ [ 'foo' ] ] ) );
+			expect( editor.getData( { trim: 'none' } ) ).to.equal( viewTable( [ [ 'foo' ] ] ) );
 
 			editor.model.change( writer => {
 				writer.remove( root.getChild( 0 ) );
 			} );
 
-			expect( editor.getData() ).to.equal( '<p>&nbsp;</p>' );
+			expect( editor.getData( { trim: 'none' } ) ).to.equal( '<p>&nbsp;</p>' );
 
 			editor.execute( 'undo' );
 
-			expect( editor.getData() ).to.equal( viewTable( [ [ 'foo' ] ] ) );
+			expect( editor.getData( { trim: 'none' } ) ).to.equal( viewTable( [ [ 'foo' ] ] ) );
 
 			editor.execute( 'redo' );
 
-			expect( editor.getData() ).to.equal( '<p>&nbsp;</p>' );
+			expect( editor.getData( { trim: 'none' } ) ).to.equal( '<p>&nbsp;</p>' );
 
 			editor.execute( 'undo' );
 
-			expect( editor.getData() ).to.equal( viewTable( [ [ 'foo' ] ] ) );
+			expect( editor.getData( { trim: 'none' } ) ).to.equal( viewTable( [ [ 'foo' ] ] ) );
 		} );
 
 		it( 'fixing empty roots should be transparent to undo - multiple roots', () => {
 			const otherRoot = doc.createRoot( '$root', 'otherRoot' );
 
-			editor.data.set( viewTable( [ [ 'foo' ] ] ), 'main' );
-			editor.data.set( viewTable( [ [ 'foo' ] ] ), 'otherRoot' );
+			editor.data.set( { main: viewTable( [ [ 'foo' ] ] ) } );
+			editor.data.set( { otherRoot: viewTable( [ [ 'foo' ] ] ) } );
 
 			editor.model.change( writer => {
 				writer.remove( root.getChild( 0 ) );
@@ -135,18 +135,18 @@ describe( 'Table feature – integration', () => {
 				writer.remove( otherRoot.getChild( 0 ) );
 			} );
 
-			expect( editor.data.get( 'main' ) ).to.equal( '<p>&nbsp;</p>' );
-			expect( editor.data.get( 'otherRoot' ) ).to.equal( '<p>&nbsp;</p>' );
+			expect( editor.data.get( { trim: 'none', rootName: 'main' } ) ).to.equal( '<p>&nbsp;</p>' );
+			expect( editor.data.get( { trim: 'none', rootName: 'otherRoot' } ) ).to.equal( '<p>&nbsp;</p>' );
 
 			editor.execute( 'undo' );
 
-			expect( editor.data.get( 'main' ) ).to.equal( '<p>&nbsp;</p>' );
-			expect( editor.data.get( 'otherRoot' ) ).to.equal( viewTable( [ [ 'foo' ] ] ) );
+			expect( editor.data.get( { trim: 'none', rootName: 'main' } ) ).to.equal( '<p>&nbsp;</p>' );
+			expect( editor.data.get( { trim: 'none', rootName: 'otherRoot' } ) ).to.equal( viewTable( [ [ 'foo' ] ] ) );
 
 			editor.execute( 'undo' );
 
-			expect( editor.data.get( 'main' ) ).to.equal( viewTable( [ [ 'foo' ] ] ) );
-			expect( editor.data.get( 'otherRoot' ) ).to.equal( viewTable( [ [ 'foo' ] ] ) );
+			expect( editor.data.get( { trim: 'none', rootName: 'main' } ) ).to.equal( viewTable( [ [ 'foo' ] ] ) );
+			expect( editor.data.get( { trim: 'none', rootName: 'otherRoot' } ) ).to.equal( viewTable( [ [ 'foo' ] ] ) );
 		} );
 	} );